+++ packages/clone-testing/src/state.rs
@@ -6,7 +6,6 @@ use cw_orch_core::{
 use cw_orch_daemon::DaemonState;
 use itertools::Itertools;
 use std::collections::HashMap;
-use tokio::runtime::Handle;
 
 #[derive(Clone, Debug)]
 /// Mock state for testing, stores addresses and code-ids.
@@ -21,13 +20,17 @@ pub struct MockState {
 
 impl MockState {
     /// Creates a new empty mock state
-    pub fn new(rt: &Handle, chain: ChainInfoOwned, deployment_id: &str) -> Self {
+    pub fn new(chain: ChainInfoOwned, deployment_id: &str) -> Self {
         Self {
             addresses: HashMap::new(),
             code_ids: HashMap::new(),
-            daemon_state: rt
-                .block_on(DaemonState::new(chain, deployment_id.to_string(), true))
-                .unwrap(),
+            daemon_state: DaemonState::new(
+                DaemonState::state_file_path().unwrap(),
+                chain,
+                deployment_id.to_string(),
+                true,
+            )
+            .unwrap(),
         }
     }
 }
@@ -92,7 +95,6 @@ mod test {
     use cw_orch_core::{environment::StateInterface, CwEnvError};
     use cw_orch_daemon::networks::JUNO_1;
     use speculoos::prelude::*;
-    use tokio::runtime::Runtime;
 
     use super::MockState;
 
@@ -100,8 +102,7 @@ mod test {
     const CONTRACT_ADDR: &str = "cosmos123";
     #[test]
     fn mock_state() {
-        let rt = Runtime::new().unwrap();
-        let mut mock = MockState::new(rt.handle(), JUNO_1.into(), "default-id");
+        let mut mock = MockState::new(JUNO_1.into(), "default-id");
         env_logger::init();
 
         let unchecked_address = &Addr::unchecked(CONTRACT_ADDR);