+++ CHANGELOG.md
@@ -11,11 +11,19 @@
 - EXCITING FEATURE : Added an item and a map query method to be able to query cw-storage-plus structure outside of contracts easily
 - Add `flush_state` method for Local Chain Daemons
 - cw-orch-interchain now errors on checking transactions for IBC packets if NO packets were found
+- `DaemonState` removed from `Sender`
+- `Channel` moved from `DaemonState` to `Sender`
+- `DaemonState` write-locks file unless it's read-only, meaning it will panic at initialization if other process holds lock of it
+- `DaemonState` now can be safely used between threads and processes
+- Two non-related Daemon's can't use same file for writing simultaneously (cloned or rebuilt are related Daemon)
+- Writing to a file happens when all Daemon's that use same file dropped instead of hot writes
+- `force_write` added to the `DaemonState` to allow force write of the state
 
 ### Breaking
 
 - Daemon : Changed return types on daemon queriers to match Cosmwasm std types
 - Cw-orch : Separate osmosis test tube from cw-orch. Its not available in its own crate `cw-orch-osmosis-test-tube`
+- Clone-testing : Remove rt in Mock State creation (daemon doesn't need it anymore)
 
 ## 0.22.0